Move waitForCompletedRequest into a generic request utils method
